user profiles set up

+++ app/controllers/likes_controller.rb
@@ -4,19 +4,21 @@ class LikesController < ApplicationController
     @bookmark = Bookmark.find(params[:bookmark_id])
     @topic = @bookmark.topic
     like = current_user.likes.build(bookmark: @bookmark)
-    
+    authorize like
+
     if like.save
       redirect_to [@topic, @bookmark], notice: 'You succesfully liked the bookmark'
     else
-      redirect_to [@topic, @bookmark], error: 'You failed to like the bookmark'      
-    end    
+      redirect_to [@topic, @bookmark], error: 'You failed to like the bookmark'
+    end
   end
-  
+
   def destroy
     @bookmark = Bookmark.find(params[:bookmark_id])
     @topic = @bookmark.topic
-    like = Like.find(params[:id])
-    
+    like = current_user.likes.find(params[:id])
+    authorize like
+
     if like.destroy
       redirect_to [@topic, @bookmark], notice: 'You unliked the bookmark'
     else
